content-generation-system/requirements.txt
python-dotenv==1.0.0
crewai==0.19.0
langchain>=0.1.10,<0.2.0
langchain_openai>=0.0.5
langchain_anthropic>=0.1.1
requests==2.31.0
streamlit==1.32.0
pydantic>=1.10.8
yaml==0.2.5
pyyaml==6.0.1
functools32>=3.2.3; python_version < '3.0'
diskcache>=5.4.0
funcy>=1.17.0